backend/requirements.txt
alembic
aniso8601
arrow
blinker
certifi
charset-normalizer
click
dateparser
Flask
Flask-Admin
Flask-Cors
Flask-DotEnv
Flask-JWT-Extended
flask-marshmallow
Flask-Migrate
Flask-RESTful
Flask-Reuploaded
Flask-SQLAlchemy
greenlet
idna
importlib-metadata
itsdangerous
Jinja2
Mako
MarkupSafe
marshmallow
marshmallow-sqlalchemy
moment
packaging
PyJWT
PyMySQL
python-dateutil
python-dotenv
pytz
regex
requests
six
SQLAlchemy
times
typing_extensions
tzlocal
urllib3
Werkzeug
WTForms
zipp 
Flask-Limiter  
uuid
datetime
enum